Removed button ERROR IN CODE, PLZ READ
The frame/GUI is not running correctly. I dont know why this is, however
when adding components to the GUI it refuses to show them. this is also
shown in the fact that there's a constant loading sign when hovering
over the window. We need to get a WORKING gui before continuing, so
that's what I"m going to work on now