Read Description
Shit tons. Removed Program.cs, as it was redundant and moved everything
to mainwindow.xaml.cs.

Got rid of the media window threading, so I also had to re-vamp the
command window.

Making big changes to config, currently bug in the Save function, #82

NOTE: there's a bunch of stuff still being synced, needs to be removed.