Creation of the separate projects, details below
I created each individual project. This will include the Desktop
Installer Project (for the creation of an installation system for the
desktop-only-version), a Module Project (for the shard's), and a server
project (for the desktop only version as well as the hearts when we port
over to that).

NOTE: The installer project is likely to be removed down the road after
we get everything set up